Add question screen, store, tests, APIs (#956)
## Reason for Pull Request

- [X] Feature/Change request

## Link for jira ticket
[CM-2178](http://evernym.atlassian.net/browse/CM-2178)

### High level description of changes done

- Added complete UI for question screen with following states of UI
  - Success (Added Lottie for success animation)
  - Error (Added Lottie for error animation)
  - Loading
  - Default status
- Added tests for UI
- Created custom transition animation while opening and closing question screen
- Used react-navigation `cardStyle` property to make background transparent for StackNavigator cards. This `cardStyle` property has to be added from top layer of StackNavigator to the bottom most layer which needs transparent card
- Utilizing stack card header to create our custom transparent header for drag animations
- Added `bg` and `color` prop to CustomText and CustomView to make them more beginner friendly. We should eventually phase our primary, secondary etc as prop name and instead move to using `bg` and `color` with values as `primary`, etc.
- Created separate `transition.js` file to handle our navigation custom transition needs
- Added three new API to bridge `signData`, `sendMessage`, `verifySignature`
- Added two new actions (`question received`, `question answered`) to be tracked for history store i.e. to show question in connection details screen
- Added new message type to process in push notification
- Added new message type to process while downloading missed messages

## Tests written for Bug/Feature/Refactoring

- `question-screen-header.spec.js`
- `question-screen.spec.js`
- `question-screen-actions.spec.js`
- `rncxs.spec.js`
- `question-store.spec.js`

## Checklist

- [X] I have created new screen, and I have checked the header of new screen on Android. I have also set the status bar color in app.js
- [X] I have checked the back button functionality required for the screens in Android. I have added the routeNames that are need to be handled for the back button in app.js
- [X] I have checked the back behaviour by swiping left or swiping down on screens and they behave as expected. Back behaviour of these screens has been approved by product team

+++ app/connection-history/connection-history-store.js
@@ -60,6 +60,11 @@ import type {
   SendProofSuccessAction,
   ProofRequestPayload,
 } from '../proof-request/type-proof-request'
+import type {
+  QuestionReceivedAction,
+  QuestionPayload,
+  UpdateQuestionAnswerAction,
+} from '../question/type-question'
 import {
   PROOF_REQUEST_RECEIVED,
   SEND_PROOF_SUCCESS,
@@ -81,6 +86,12 @@ import { RESET } from '../common/type-common'
 import { CLAIM_OFFER_RECEIVED } from '../claim-offer/type-claim-offer'
 import { captureError } from '../services/error/error-handler'
 import { customLogger } from '../store/custom-logger'
+import {
+  QUESTION_RECEIVED,
+  UPDATE_QUESTION_ANSWER,
+} from '../question/type-question'
+import { MESSAGE_TYPE } from '../api/api-constants'
+import { selectQuestion } from '../question/question-store'
 
 const initialState = {
   error: null,
@@ -294,6 +305,45 @@ export function convertProofSendToHistoryEvent(
   }
 }
 
+export function convertQuestionReceivedToHistoryEvent(
+  action: QuestionReceivedAction
+): ConnectionHistoryEvent {
+  return {
+    action: HISTORY_EVENT_STATUS[QUESTION_RECEIVED],
+    data: action.question,
+    id: uuid(),
+    name: action.question.messageTitle,
+    status: HISTORY_EVENT_STATUS[QUESTION_RECEIVED],
+    timestamp: moment().format(),
+    type: HISTORY_EVENT_TYPE.QUESTION,
+    remoteDid: action.question.from_did,
+    originalPayload: {
+      payloadInfo: action.question,
+      type: MESSAGE_TYPE.QUESTION,
+    },
+  }
+}
+
+export function convertQuestionAnswerToHistoryEvent(
+  action: UpdateQuestionAnswerAction,
+  question: QuestionPayload
+): ConnectionHistoryEvent {
+  return {
+    action: HISTORY_EVENT_STATUS[UPDATE_QUESTION_ANSWER],
+    data: { payload: question, ...action },
+    id: uuid(),
+    name: `You responded with: ${action.answer.text}`,
+    status: HISTORY_EVENT_STATUS[UPDATE_QUESTION_ANSWER],
+    timestamp: moment().format(),
+    type: HISTORY_EVENT_TYPE.QUESTION,
+    remoteDid: question.from_did,
+    originalPayload: {
+      payloadInfo: question,
+      type: MESSAGE_TYPE.QUESTION,
+    },
+  }
+}
+
 export const recordHistoryEvent = (historyEvent: ConnectionHistoryEvent) => ({
   type: RECORD_HISTORY_EVENT,
   historyEvent,
@@ -398,6 +448,25 @@ export function* historyEventOccurredSaga(
         historyEvent.originalPayload.uid,
         historyEvent.remoteDid,
         PROOF_REQUEST_RECEIVED
+      )
+      if (oldHistoryEvent) yield put(deleteHistoryEvent(oldHistoryEvent))
+    }
+
+    if (event.type === QUESTION_RECEIVED) {
+      historyEvent = convertQuestionReceivedToHistoryEvent(event)
+    }
+
+    if (event.type === UPDATE_QUESTION_ANSWER) {
+      const questionPayload: QuestionPayload = yield select(
+        selectQuestion,
+        event.uid
+      )
+      historyEvent = convertQuestionAnswerToHistoryEvent(event, questionPayload)
+      const oldHistoryEvent = yield select(
+        getHistoryEvent,
+        event.uid,
+        historyEvent.remoteDid,
+        MESSAGE_TYPE.QUESTION
       )
       if (oldHistoryEvent) yield put(deleteHistoryEvent(oldHistoryEvent))
     }
